Make the deprecated values more clear
At first I read it as though you should use `"true"` or `"false"`, but upon looking closer it seems that it should be `true` and `false`. This change should disambiguate it better. Also, the final `.` doesn't seem necessary, there is now a double period at the end of the line in the actual documentation.